requirements.txt
telebot~=0.0.5
tensorflow~=2.14.0
Pillow~=10.1.0
opencv-python~=4.8.1.78
numpy~=1.26.1
